Split watch data classes into different files Added time display to detail view of lesson (#7)
